split up scripts
