+++ feed/index.html
@@ -74,17 +74,21 @@
       <h2 class="text-frontpage font-fm-mulish border-top border-warning w-75 m-auto fs-1-5-rem pt-3">Other users looking for a Hikebuddy</h2>
     </div>
     <div class="row mt-3 justify-content-center mx-0">
-      <div class="col-8 col-sm-4 col-md-2">
+      <div class="col-6 col-sm-4 col-md-2">
       <select id="sortOrderSelect" class="form-select form-select-sm p-1" aria-label="Sort user posts by">
         <option selected value="desc">Newest</option>
         <option value="asc">Oldest</option>
       </select>
+    </div>
+      <div class="col-6 col-sm-4 col-md-2">
+      <select id="filter-from-tags" class="form-select form-select-sm p-1" aria-label="Filter posts by tags">
+        <option selected value="desc">Sort by tags</option>
+      </select>
+    </div>
 
+      </div>
       <div id="spinner-feed" class="spinner-border text-warning mt-5" role="status">
         <span class="sr-only"></span>
-      </div>
-
-    </div>
     </div>
     <div class="feed-container row justify-content-center mt-3 mx-0">
     </div>
