+++ js/validation/validate-email.js
@@ -1,5 +1,3 @@
-
-
 export function validateEmail(regEmail) {
   const regEx = /^[a-zA-Z0-9._%+-]+@(noroff\.no|stud\.noroff\.no)$/;
   const patternMatches = regEx.test(regEmail);
