+++ js/validation/validate-post-form.mjs
@@ -1,0 +1,7 @@
+export function validatePostForm() {
+  if (!messageTitle.value.trim() || !messageContent.value.trim()) {
+    alert("Both title and content are required. This will be changed lager");
+    return false;
+  } 
+  return true
+  }
