+++ src/scss/index.scss
@@ -16,6 +16,13 @@
 
 @import "buttons";
 
+// Container styling
+
+.container {
+  word-wrap: break-word !important;
+  overflow-wrap: break-word !important;
+ }
+
 // Border-radius
 
   // Border-radius variables
